% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambra_functions.R
\name{append_patient_df}
\alias{append_patient_df}
\title{Function to append patient data to master dataframe based on column name}
\usage{
append_patient_df(patient_df, master_df)
}
\arguments{
\item{patient_df}{Patient dataframe generated by \code{\link{retrieve_patient_attachment}}}

\item{master_df}{Master dataframe for a given account.}
}
\value{
The master dataframe with the patient row appended.
}
\description{
Function to append patient data to master dataframe based on column name
}
\examples{
master_df <- append_patient_df(patient_df, master_df)
}
