% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambra_functions.R
\name{get_all_patients}
\alias{get_all_patients}
\title{A function to get all patients for a group or location.}
\usage{
get_all_patients(group_or_locations_names_and_ids)
}
\arguments{
\item{group_or_locations_names_and_ids}{Ambra generated python iterable for all locations or groups in a namespace.}
}
\value{
A list of all patients with patientid, engine_fqdn, storage_namespace, and study_uid.
}
\description{
A function to get all patients for a group or location.
}
\examples{
all_patients <- get_all_patients(group_or_locations_names_and_ids)
}
