% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambra_functions.R
\name{get_completed_patients}
\alias{get_completed_patients}
\title{Retrieves all completed patients for a given account.}
\usage{
get_completed_patients(group_names_and_ids)
}
\arguments{
\item{group_names_and_ids}{A list of lists, generated from \code{\link{get_group_names_and_ids}}.}
}
\value{
A list of all completed patients with patientid, engine_fqdn, storage_namespace, study_uid, and attachment count.
}
\description{
Retrieves all completed patients for a given account.
}
\examples{
completed_patients <- get_completed_patients(group_names_and_ids)
}
