% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambra_functions.R
\name{get_group_names_and_ids}
\alias{get_group_names_and_ids}
\title{Function to retrieve all group names and ids for a given account namespace.}
\usage{
get_group_names_and_ids(group_iterable)
}
\arguments{
\item{group_iterable}{Ambra generated python iterable for all groups in a namespace.}
}
\value{
A list with the name and namespace id of each group in the account namespace.
}
\description{
Function to retrieve all group names and ids for a given account namespace.
}
\examples{
api <- ambrasdk$api$Api$with_creds(url = 'https://access.dicomgrid.com/api/v3', username = 'XXX', password = 'XXX')
account <- api$Account$list()$filter_by(Ambra_filter('name', Ambra_filter_cond$equals, account_name))$first()
groups <- api$Group$list(account_id = account$uuid)$all()
group_names_and_ids <- get_group_names_and_ids(groups)
}
