% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambra_functions.R
\name{get_location_names_and_ids}
\alias{get_location_names_and_ids}
\title{Function to retrieve all location names and ids for a given account namespace.}
\usage{
get_location_names_and_ids(location_iterable)
}
\arguments{
\item{location_iterable}{Ambra generated python iterable for all locations in a namespace.}
}
\value{
A list with the name and namespace id of each location in the account namespace.
}
\description{
Function to retrieve all location names and ids for a given account namespace.
}
\examples{
location_names_and_ids <- get_location_names_and_ids(locations)
}
