% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambra_functions.R
\name{get_total_patients_per_location}
\alias{get_total_patients_per_location}
\title{Function to get total patients per location from AMBRA.}
\usage{
get_total_patients_per_location(all_location_names_and_ids, location_list)
}
\arguments{
\item{all_location_names_and_ids}{List of all locations and their namespace ids.}

\item{location_list}{List of locations of interest to filter all locations list.}
}
\value{
A list of each location and how many unique patients per location.
}
\description{
Function to get total patients per location from AMBRA.
}
\examples{
site_numbers <- get_total_patients_per_location(location_names_and_ids, site_locations)
}
