% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambra_functions.R
\name{retrieve_patient_attachment}
\alias{retrieve_patient_attachment}
\title{Retrieves latest attachment from AMBRA and converts it into a data frame.}
\usage{
retrieve_patient_attachment(patient_schema)
}
\arguments{
\item{patient_schema}{An entry from the list of lists generated from \code{\link{get_completed_patients}}.}
}
\value{
A dataframe with the eCRF data for the given patient.
}
\description{
Retrieves latest attachment from AMBRA and converts it into a data frame.
}
\examples{
patient_df <- retrieve_patient_attachment(completed_patients[i]) #where i is an index
patient_df <- retrieve_patient_attachment(patient_schema)
}
