% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambra_functions.R
\name{update_master_spreadsheet}
\alias{update_master_spreadsheet}
\title{A function to update and save master spreadsheet with all completed patients.}
\usage{
update_master_spreadsheet(path_to_master_spreadsheet, completed_patients)
}
\arguments{
\item{path_to_master_spreadsheet}{Absolute or relative path to master spreadsheet for given account.}

\item{completed_patients}{List of completed patients generated from \code{\link{get_completed_patients}}}
}
\value{
Nothing is returned, automatically saves updated spreadsheet to path_to_master_spreadsheet.
}
\description{
A function to update and save master spreadsheet with all completed patients.
}
\examples{
update_master_spreadsheet(path_to_master_spreadsheet, completed_patients)
}
